pyproject.toml
[project]
name = "cpcconformertransfomerspeakerdiarizationmodel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cpc-audio",
    "huggingface-hub[cli]>=0.34.4",
    "numpy>=2.3.2",
    "scipy>=1.16.1",
    "soundfile>=0.13.1",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.12",
]

[tool.uv.sources]
cpc-audio = { git = "https://github.com/facebookresearch/CPC_audio.git", rev = "b98a1bdf1fe9ea219816db7a6c28115d404a3510" }
